BugAssigner/requirements.txt
Django==3.1.7
django-allauth==0.45.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-embed-video==1.4.0
django-localflavor==3.0.1
django-memcache-status==2.2
django-parler==2.2
django-rest-auth==0.9.5
django-rosetta==0.9.5
django-taggit==1.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
easy-thumbnails==2.7.1
keras==2.7.0
Keras-Preprocessing==1.1.2
nltk==3.6.6
numpy==1.21.5
pandas==1.3.5
Pillow==8.1.2
requests==2.25.1
requests-oauthlib==1.3.0
tensorflow==2.7.0
